move observeCanUploadFiles and observeCanDownloadFiles to queries/security
