+++ app/components/draft_scheduled_post/draft_scheduled_post_container/draft_and_scheduled_post_files/index.ts
@@ -4,7 +4,8 @@
 import {withDatabase, withObservables} from '@nozbe/watermelondb/react';
 
 import Files from '@components/files/files';
-import {observeCanDownloadFiles, observeConfigBooleanValue} from '@queries/servers/system';
+import {observeCanDownloadFiles} from '@queries/servers/security';
+import {observeConfigBooleanValue} from '@queries/servers/system';
 
 import type {WithDatabaseArgs} from '@typings/database/database';
